+++ .eslintrc.js
@@ -1,23 +1,23 @@
 module.exports = {
-    extends: ['@zazen/eslint-config'],
-    plugins: ['@typescript-eslint'],
-    rules: {},
+    extends: ['@zazen/eslint-config/typescript'],
+    env: {
+        node: true,
+    },
+    rules: {
+        '@typescript-eslint/indent': 'off',
+        '@typescript-eslint/space-before-function-paren': 'off',
+
+        '@typescript-eslint/explicit-function-return-type': 'off',
+    },
     overrides: [
         {
-            files: ['**/*.ts'],
-            extends: [
-                '@zazen/eslint-config',
-                // https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/src/configs/recommended.json
-                'plugin:@typescript-eslint/recommended',
-                // https://github.com/prettier/eslint-config-prettier/blob/master/%40typescript-eslint.js
-                'prettier/@typescript-eslint',
+            // Jest config
+            files: [
+                '**/__tests__/**/*.{js,ts,tsx}',
+                '**/*.@(spec|test).{js,ts,tsx}',
             ],
-            parser: '@typescript-eslint/parser',
             env: {
-                node: true,
-            },
-            rules: {
-                'prefer-const': 'off',
+                jest: true,
             },
         },
     ],