+++ __tests__/plugin.test.ts
@@ -1,7 +1,11 @@
+import { diffStringsUnified } from 'jest-diff'
 import postcss from 'postcss'
+import { format } from 'prettier'
 import tailwindcss from 'tailwindcss'
 
-import capsizePlugin from '../dist'
+// Need to test this how consumers will use it, using `require()`.
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const capsizePlugin = require('../dist')
 
 const THEME_CONFIG = {
     fontFamily: {
@@ -19,18 +23,25 @@ const THEME_CONFIG = {
 }
 
 expect.extend({
-    toMatchCss: (received, argument) => {
-        let stripped = (str: string) => str.replace(/[;\s]/g, '')
+    toMatchCss: (receivedCss: string, expectedCss: string) => {
+        let strip = (str: string) => str.replace(/[;\s]/g, '')
 
-        if (stripped(received) === stripped(argument)) {
+        if (strip(receivedCss) === strip(expectedCss)) {
             return {
                 message: () =>
-                    `expected ${received} not to match CSS ${argument}`,
+                    `expected ${receivedCss} not to match CSS ${expectedCss}`,
                 pass: true,
             }
         } else {
+            let receivedCssFormatted = format(receivedCss, { parser: 'css' })
+            let expectedCssFormatted = format(expectedCss, { parser: 'css' })
+            let diff = diffStringsUnified(
+                receivedCssFormatted,
+                expectedCssFormatted,
+            )
+
             return {
-                message: () => `expected ${received} to match CSS ${argument}`,
+                message: () => `expected CSS to match:\n${diff}`,
                 pass: false,
             }
         }
@@ -39,7 +50,7 @@ expect.extend({
 
 describe('Plugin', () => {
     it('generates utility classes with a default root size', async () => {
-        return postcss(
+        return await postcss(
             tailwindcss({
                 theme: {
                     ...THEME_CONFIG,
@@ -60,26 +71,15 @@ describe('Plugin', () => {
                 from: undefined,
             })
             .then((result) => {
-                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                // @ts-ignore
                 expect(result.css).toMatchCss(`
-                    .font-sans.text-sm.leading-sm.capsize,
-                    .font-sans .text-sm.leading-sm.capsize,
-                    .font-sans .text-sm .leading-sm.capsize,
-                    .text-sm .font-sans.leading-sm.capsize,
-                    .text-sm .font-sans .leading-sm.capsize {
-                        padding: 0.05px 0;
-                    }
-
                     .font-sans.text-sm.leading-sm.capsize::before,
                     .font-sans .text-sm.leading-sm.capsize::before,
                     .font-sans .text-sm .leading-sm.capsize::before,
                     .text-sm .font-sans.leading-sm.capsize::before,
                     .text-sm .font-sans .leading-sm.capsize::before {
                         content: '';
-                        margin-top: -0.3542em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -0.3506em;
+                        display: table;
                     }
 
                     .font-sans.text-sm.leading-sm.capsize::after,
@@ -88,17 +88,8 @@ describe('Plugin', () => {
                     .text-sm .font-sans.leading-sm.capsize::after,
                     .text-sm .font-sans .leading-sm.capsize::after {
                         content: '';
-                        margin-bottom: -0.3542em;
-                        display: block;
-                        height: 0;
-                    }
-
-                    .font-sans.text-sm.leading-md.capsize,
-                    .font-sans .text-sm.leading-md.capsize,
-                    .font-sans .text-sm .leading-md.capsize,
-                    .text-sm .font-sans.leading-md.capsize,
-                    .text-sm .font-sans .leading-md.capsize {
-                        padding: 0.05px 0;
+                        margin-top: -0.3506em;
+                        display: table;
                     }
 
                     .font-sans.text-sm.leading-md.capsize::before,
@@ -107,9 +98,8 @@ describe('Plugin', () => {
                     .text-sm .font-sans.leading-md.capsize::before,
                     .text-sm .font-sans .leading-md.capsize::before {
                         content: '';
-                        margin-top: -1.0685em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -1.0649em;
+                        display: table;
                     }
 
                     .font-sans.text-sm.leading-md.capsize::after,
@@ -118,17 +108,8 @@ describe('Plugin', () => {
                     .text-sm .font-sans.leading-md.capsize::after,
                     .text-sm .font-sans .leading-md.capsize::after {
                         content: '';
-                        margin-bottom: -1.0685em;
-                        display: block;
-                        height: 0;
-                    }
-
-                    .font-sans.text-md.leading-sm.capsize,
-                    .font-sans .text-md.leading-sm.capsize,
-                    .font-sans .text-md .leading-sm.capsize,
-                    .text-md .font-sans.leading-sm.capsize,
-                    .text-md .font-sans .leading-sm.capsize {
-                        padding: 0.05px 0;
+                        margin-top: -1.0649em;
+                        display: table;
                     }
 
                     .font-sans.text-md.leading-sm.capsize::before,
@@ -137,9 +118,8 @@ describe('Plugin', () => {
                     .text-md .font-sans.leading-sm.capsize::before,
                     .text-md .font-sans .leading-sm.capsize::before {
                         content: '';
-                        margin-top: -0.0551em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -0.053em;
+                        display: table;
                     }
 
                     .font-sans.text-md.leading-sm.capsize::after,
@@ -148,17 +128,8 @@ describe('Plugin', () => {
                     .text-md .font-sans.leading-sm.capsize::after,
                     .text-md .font-sans .leading-sm.capsize::after {
                         content: '';
-                        margin-bottom: -0.0551em;
-                        display: block;
-                        height: 0;
-                    }
-
-                    .font-sans.text-md.leading-md.capsize,
-                    .font-sans .text-md.leading-md.capsize,
-                    .font-sans .text-md .leading-md.capsize,
-                    .text-md .font-sans.leading-md.capsize,
-                    .text-md .font-sans .leading-md.capsize {
-                        padding: 0.05px 0;
+                        margin-top: -0.053em;
+                        display: table;
                     }
 
                     .font-sans.text-md.leading-md.capsize::before,
@@ -167,9 +138,8 @@ describe('Plugin', () => {
                     .text-md .font-sans.leading-md.capsize::before,
                     .text-md .font-sans .leading-md.capsize::before {
                         content: '';
-                        margin-top: -0.4718em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -0.4697em;
+                        display: table;
                     }
 
                     .font-sans.text-md.leading-md.capsize::after,
@@ -178,16 +148,15 @@ describe('Plugin', () => {
                     .text-md .font-sans.leading-md.capsize::after,
                     .text-md .font-sans .leading-md.capsize::after {
                         content: '';
-                        margin-bottom: -0.4718em;
-                        display: block;
-                        height: 0;
+                        margin-top: -0.4697em;
+                        display: table;
                     }
                 `)
             })
     })
 
     it('generates utility classes with a custom root size', async () => {
-        return postcss(
+        return await postcss(
             tailwindcss({
                 theme: {
                     ...THEME_CONFIG,
@@ -208,26 +177,15 @@ describe('Plugin', () => {
                 from: undefined,
             })
             .then((result) => {
-                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                // @ts-ignore
                 expect(result.css).toMatchCss(`
-                    .font-sans.text-sm.leading-sm.capsize,
-                    .font-sans .text-sm.leading-sm.capsize,
-                    .font-sans .text-sm .leading-sm.capsize,
-                    .text-sm .font-sans.leading-sm.capsize,
-                    .text-sm .font-sans .leading-sm.capsize {
-                        padding: 0.05px 0;
-                    }
-
                     .font-sans.text-sm.leading-sm.capsize::before,
                     .font-sans .text-sm.leading-sm.capsize::before,
                     .font-sans .text-sm .leading-sm.capsize::before,
                     .text-sm .font-sans.leading-sm.capsize::before,
                     .text-sm .font-sans .leading-sm.capsize::before {
                         content: '';
-                        margin-top: -0.3542em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -0.3506em;
+                        display: table;
                     }
 
                     .font-sans.text-sm.leading-sm.capsize::after,
@@ -236,17 +194,8 @@ describe('Plugin', () => {
                     .text-sm .font-sans.leading-sm.capsize::after,
                     .text-sm .font-sans .leading-sm.capsize::after {
                         content: '';
-                        margin-bottom: -0.3542em;
-                        display: block;
-                        height: 0;
-                    }
-
-                    .font-sans.text-sm.leading-md.capsize,
-                    .font-sans .text-sm.leading-md.capsize,
-                    .font-sans .text-sm .leading-md.capsize,
-                    .text-sm .font-sans.leading-md.capsize,
-                    .text-sm .font-sans .leading-md.capsize {
-                        padding: 0.05px 0;
+                        margin-top: -0.3506em;
+                        display: table;
                     }
 
                     .font-sans.text-sm.leading-md.capsize::before,
@@ -255,9 +204,8 @@ describe('Plugin', () => {
                     .text-sm .font-sans.leading-md.capsize::before,
                     .text-sm .font-sans .leading-md.capsize::before {
                         content: '';
-                        margin-top: -0.7114em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -0.7078em;
+                        display: table;
                     }
 
                     .font-sans.text-sm.leading-md.capsize::after,
@@ -266,17 +214,8 @@ describe('Plugin', () => {
                     .text-sm .font-sans.leading-md.capsize::after,
                     .text-sm .font-sans .leading-md.capsize::after {
                         content: '';
-                        margin-bottom: -0.7114em;
-                        display: block;
-                        height: 0;
-                    }
-
-                    .font-sans.text-md.leading-sm.capsize,
-                    .font-sans .text-md.leading-sm.capsize,
-                    .font-sans .text-md .leading-sm.capsize,
-                    .text-md .font-sans.leading-sm.capsize,
-                    .text-md .font-sans .leading-sm.capsize {
-                        padding: 0.05px 0;
+                        margin-top: -0.7078em;
+                        display: table;
                     }
 
                     .font-sans.text-md.leading-sm.capsize::before,
@@ -285,9 +224,8 @@ describe('Plugin', () => {
                     .text-md .font-sans.leading-sm.capsize::before,
                     .text-md .font-sans .leading-sm.capsize::before {
                         content: '';
-                        margin-top: -0.1947em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -0.1919em;
+                        display: table;
                     }
 
                     .font-sans.text-md.leading-sm.capsize::after,
@@ -296,17 +234,8 @@ describe('Plugin', () => {
                     .text-md .font-sans.leading-sm.capsize::after,
                     .text-md .font-sans .leading-sm.capsize::after {
                         content: '';
-                        margin-bottom: -0.1947em;
-                        display: block;
-                        height: 0;
-                    }
-
-                    .font-sans.text-md.leading-md.capsize,
-                    .font-sans .text-md.leading-md.capsize,
-                    .font-sans .text-md .leading-md.capsize,
-                    .text-md .font-sans.leading-md.capsize,
-                    .text-md .font-sans .leading-md.capsize {
-                        padding: 0.05px 0;
+                        margin-top: -0.1919em;
+                        display: table;
                     }
 
                     .font-sans.text-md.leading-md.capsize::before,
@@ -315,9 +244,8 @@ describe('Plugin', () => {
                     .text-md .font-sans.leading-md.capsize::before,
                     .text-md .font-sans .leading-md.capsize::before {
                         content: '';
-                        margin-top: -0.4725em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -0.4697em;
+                        display: table;
                     }
 
                     .font-sans.text-md.leading-md.capsize::after,
@@ -326,16 +254,15 @@ describe('Plugin', () => {
                     .text-md .font-sans.leading-md.capsize::after,
                     .text-md .font-sans .leading-md.capsize::after {
                         content: '';
-                        margin-bottom: -0.4725em;
-                        display: block;
-                        height: 0;
+                        margin-top: -0.4697em;
+                        display: table;
                     }
                 `)
             })
     })
 
     it('works with unitless or percentage line-height values', async () => {
-        return postcss(
+        return await postcss(
             tailwindcss({
                 theme: {
                     ...THEME_CONFIG,
@@ -355,26 +282,15 @@ describe('Plugin', () => {
                 from: undefined,
             })
             .then((result) => {
-                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                // @ts-ignore
                 expect(result.css).toMatchCss(`
-                    .font-sans.text-sm.leading-sm.capsize,
-                    .font-sans .text-sm.leading-sm.capsize,
-                    .font-sans .text-sm .leading-sm.capsize,
-                    .text-sm .font-sans.leading-sm.capsize,
-                    .text-sm .font-sans .leading-sm.capsize {
-                        padding: 0.05px 0;
-                    }
-
                     .font-sans.text-sm.leading-sm.capsize::before,
                     .font-sans .text-sm.leading-sm.capsize::before,
                     .font-sans .text-sm .leading-sm.capsize::before,
                     .text-sm .font-sans.leading-sm.capsize::before,
                     .text-sm .font-sans .leading-sm.capsize::before {
                         content: '';
-                        margin-top: -0.1395em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -0.1364em;
+                        display: table;
                     }
 
                     .font-sans.text-sm.leading-sm.capsize::after,
@@ -383,17 +299,8 @@ describe('Plugin', () => {
                     .text-sm .font-sans.leading-sm.capsize::after,
                     .text-sm .font-sans .leading-sm.capsize::after {
                         content: '';
-                        margin-bottom: -0.1395em;
-                        display: block;
-                        height: 0;
-                    }
-
-                    .font-sans.text-sm.leading-md.capsize,
-                    .font-sans .text-sm.leading-md.capsize,
-                    .font-sans .text-sm .leading-md.capsize,
-                    .text-sm .font-sans.leading-md.capsize,
-                    .text-sm .font-sans .leading-md.capsize {
-                        padding: 0.05px 0;
+                        margin-top: -0.1364em;
+                        display: table;
                     }
 
                     .font-sans.text-sm.leading-md.capsize::before,
@@ -402,9 +309,8 @@ describe('Plugin', () => {
                     .text-sm .font-sans.leading-md.capsize::before,
                     .text-sm .font-sans .leading-md.capsize::before {
                         content: '';
-                        margin-top: -0.3895em;
-                        display: block;
-                        height: 0;
+                        margin-bottom: -0.3864em;
+                        display: table;
                     }
 
                     .font-sans.text-sm.leading-md.capsize::after,
@@ -413,9 +319,52 @@ describe('Plugin', () => {
                     .text-sm .font-sans.leading-md.capsize::after,
                     .text-sm .font-sans .leading-md.capsize::after {
                         content: '';
-                        margin-bottom: -0.3895em;
-                        display: block;
-                        height: 0;
+                        margin-top: -0.3864em;
+                        display: table;
+                    }
+                `)
+            })
+    })
+
+    it('generates utility classes with a custom activation class', async () => {
+        return await postcss(
+            tailwindcss({
+                theme: {
+                    ...THEME_CONFIG,
+                    fontSize: {
+                        sm: '1rem',
+                    },
+                    lineHeight: {
+                        md: '1.5',
+                    },
+                },
+                corePlugins: false,
+                plugins: [capsizePlugin({ className: 'leading-trim' })],
+            }),
+        )
+            .process('@tailwind components; @tailwind utilities', {
+                from: undefined,
+            })
+            .then((result) => {
+                expect(result.css).toMatchCss(`
+                    .font-sans.text-sm.leading-md.leading-trim::before,
+                    .font-sans .text-sm.leading-md.leading-trim::before,
+                    .font-sans .text-sm .leading-md.leading-trim::before,
+                    .text-sm .font-sans.leading-md.leading-trim::before,
+                    .text-sm .font-sans .leading-md.leading-trim::before {
+                        content: '';
+                        margin-bottom: -0.3864em;
+                        display: table;
+                    }
+
+                    .font-sans.text-sm.leading-md.leading-trim::after,
+                    .font-sans .text-sm.leading-md.leading-trim::after,
+                    .font-sans .text-sm .leading-md.leading-trim::after,
+                    .text-sm .font-sans.leading-md.leading-trim::after,
+                    .text-sm .font-sans .leading-md.leading-trim::after {
+                        content: '';
+                        margin-top: -0.3864em;
+                        display: table;
                     }
                 `)
             })